Envio de email Async

+++ src/main/java/br/com/ifpe/matafome_api/api/produto/ProdutoRequest.java
@@ -19,7 +19,7 @@ public class ProdutoRequest {
 
     private String nome;
 
-    private Double preco;
+    private Float preco;
 
     private String descricao;
