+++ src/main/java/br/com/ifpe/matafome_api/modelo/produto/Produto.java
@@ -36,7 +36,7 @@ public class Produto extends EntidadeAuditavel {
     private String nome;
 
     @Column
-    private Double preco;
+    private Float preco;
 
     @Column
     private String descricao;
